\name{hclda-internal}
\title{Internal hclda Functions}
\alias{generate_data_hclda}
\alias{creasteSS_Rcpp}
\alias{computeY_Rcpp}
\alias{creasteSS_fast_Rcpp}
\alias{loo_CLD_Rcpp}
\alias{fast_loo_CLD_Rcpp}
\alias{cv_Rcpp}
\alias{cv_clustering_Rcpp}
\alias{computeA}
\alias{parhclda}
\description{Internal hclda Functions}
%\usage{
%Cov_beta(gam, xx, beta)
%}
\author{Kanta Miura and Kei Hirose}
%\details{
%aaa
%}
\keyword{internal}
